Restructure `worker` module
This contains no behavior changes. It only shuffles some of the code around into different modules. In this new module layout the `worker` module contains all code related to our background job processing. The `worker::swirl` submodule contains the code for the generic background job runner, and the `worker::jobs` submodule contains the application-specific background job definitions.
