+++ src/admin/delete_crate.rs
@@ -1,5 +1,5 @@
-use crate::background_jobs::enqueue_sync_to_index;
 use crate::storage::Storage;
+use crate::worker::jobs;
 use crate::{admin::dialoguer, db, schema::crates};
 use anyhow::Context;
 use diesel::prelude::*;
@@ -71,7 +71,7 @@ pub fn run(opts: Opts) {
         };
 
         info!(%name, "Enqueuing index sync jobs");
-        if let Err(error) = enqueue_sync_to_index(name, conn) {
+        if let Err(error) = jobs::enqueue_sync_to_index(name, conn) {
             warn!(%name, ?error, "Failed to enqueue index sync jobs");
         }
