+++ src/admin/enqueue_job.rs
@@ -1,9 +1,7 @@
-use crate::background_jobs::BackgroundJob;
 use crate::db;
 use crate::schema::background_jobs;
-use crate::worker::{
-    DailyDbMaintenanceJob, DumpDbJob, NormalizeIndexJob, SquashIndexJob, UpdateDownloadsJob,
-};
+use crate::worker::jobs;
+use crate::worker::swirl::BackgroundJob;
 use anyhow::Result;
 use diesel::prelude::*;
 use secrecy::{ExposeSecret, SecretString};
@@ -46,15 +44,17 @@ pub fn run(command: Command) -> Result<()> {
                 println!("Did not enqueue update_downloads, existing job already in progress");
                 Ok(())
             } else {
-                Ok(UpdateDownloadsJob.enqueue(conn)?)
+                Ok(jobs::UpdateDownloads.enqueue(conn)?)
             }
         }
         Command::DumpDb {
             database_url,
             target_name,
-        } => Ok(DumpDbJob::new(database_url.expose_secret(), target_name).enqueue(conn)?),
-        Command::DailyDbMaintenance => Ok(DailyDbMaintenanceJob.enqueue(conn)?),
-        Command::SquashIndex => Ok(SquashIndexJob.enqueue(conn)?),
-        Command::NormalizeIndex { dry_run } => Ok(NormalizeIndexJob::new(dry_run).enqueue(conn)?),
+        } => Ok(jobs::DumpDb::new(database_url.expose_secret(), target_name).enqueue(conn)?),
+        Command::DailyDbMaintenance => Ok(jobs::DailyDbMaintenance.enqueue(conn)?),
+        Command::SquashIndex => Ok(jobs::SquashIndex.enqueue(conn)?),
+        Command::NormalizeIndex { dry_run } => {
+            Ok(jobs::NormalizeIndex::new(dry_run).enqueue(conn)?)
+        }
     }
 }
