--- src/swirl.rs
@@ -1,7 +1,0 @@
-mod runner;
-mod storage;
-
-pub mod errors;
-
-pub use self::runner::Runner;
-pub(crate) use errors::PerformError;
