+++ src/worker/mod.rs
@@ -1,25 +1,18 @@
-//! The `worker` module contains all the tasks that can be queued up for the
-//! background worker process to work on. This includes recurring tasks like
-//! the daily database maintenance, but also operations like rendering READMEs
-//! and uploading them to S3.
+//! This module contains the code for the background jobs that run on the
+//! crates.io backend servers.
+//!
+//! The `swirl` submodule contains the code for the generic background job
+//! runner, and the `jobs` submodule contains the application-specific
+//! background job definitions.
 
-pub mod cloudfront;
-mod daily_db_maintenance;
-pub mod dump_db;
-pub mod fastly;
-mod git;
-mod readmes;
-mod update_downloads;
-
-pub(crate) use daily_db_maintenance::DailyDbMaintenanceJob;
-pub(crate) use dump_db::DumpDbJob;
-pub(crate) use git::{NormalizeIndexJob, SquashIndexJob, SyncToGitIndexJob, SyncToSparseIndexJob};
-pub(crate) use readmes::RenderAndUploadReadmeJob;
-pub(crate) use update_downloads::UpdateDownloadsJob;
-
-use crate::background_jobs::Environment;
-use crate::swirl::Runner;
+use self::swirl::Runner;
 use std::sync::Arc;
+
+mod environment;
+pub mod jobs;
+pub mod swirl;
+
+pub use self::environment::Environment;
 
 pub trait RunnerExt {
     fn register_crates_io_job_types(self) -> Self;
@@ -27,13 +20,13 @@ pub trait RunnerExt {
 
 impl RunnerExt for Runner<Arc<Environment>> {
     fn register_crates_io_job_types(self) -> Self {
-        self.register_job_type::<DailyDbMaintenanceJob>()
-            .register_job_type::<DumpDbJob>()
-            .register_job_type::<NormalizeIndexJob>()
-            .register_job_type::<RenderAndUploadReadmeJob>()
-            .register_job_type::<SquashIndexJob>()
-            .register_job_type::<SyncToGitIndexJob>()
-            .register_job_type::<SyncToSparseIndexJob>()
-            .register_job_type::<UpdateDownloadsJob>()
+        self.register_job_type::<jobs::DailyDbMaintenance>()
+            .register_job_type::<jobs::DumpDb>()
+            .register_job_type::<jobs::NormalizeIndex>()
+            .register_job_type::<jobs::RenderAndUploadReadme>()
+            .register_job_type::<jobs::SquashIndex>()
+            .register_job_type::<jobs::SyncToGitIndex>()
+            .register_job_type::<jobs::SyncToSparseIndex>()
+            .register_job_type::<jobs::UpdateDownloads>()
     }
 }
